added auto zipping and running of beatmap

+++ osu!MapRandomizer/Form1.cs
@@ -10,6 +10,7 @@ using System.Windows.Forms;
 using System.IO;
 using System.Diagnostics;
 using System.Threading;
+using System.IO.Compression;
 using OsuParsers.Beatmaps;
 using OsuParsers.Decoders;
 
@@ -64,6 +65,14 @@ namespace osu_MapRandomizer
         private void button2_Click(object sender, EventArgs e)
         {
             RandomizeBeatmap();
+            try
+            {
+                
+            }
+            catch
+            {
+
+            }
         }
 
         void Log(string text)
@@ -103,7 +112,16 @@ namespace osu_MapRandomizer
 
         void RandomizeBeatmap()
         {
-            Directory.Delete("RandomizedMap", true);
+            if (Directory.Exists("RandomizedMap"))
+            {
+                Directory.Delete("RandomizedMap", true);
+            }
+
+            if (File.Exists("RandomizedMap.osz"))
+            {
+                File.Delete("RandomizedMap.osz");
+            }
+
             Random r = new Random();
             var kmap = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
             var ksong = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
@@ -115,20 +133,80 @@ namespace osu_MapRandomizer
                 kmap = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
                 ksong = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
             }
+            //create map directory
+            Directory.CreateDirectory("RandomizedMap");
 
             map.TimingPoints = song.TimingPoints;
             map.GeneralSection.AudioFilename = song.GeneralSection.AudioFilename;
             map.GeneralSection.Length = song.GeneralSection.Length;
+            map.EventsSection.Breaks = song.EventsSection.Breaks;
+
+            //copy audio file
+            File.Copy(Path.Combine(Path.GetDirectoryName(ksong.FullName), song.GeneralSection.AudioFilename), Path.Combine("RandomizedMap", song.GeneralSection.AudioFilename));
+
+            //video randomizer
+            if (checkedListBox1.GetItemChecked(2))
+            {
+                var kvid = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
+                Beatmap vid = BeatmapDecoder.Decode(kvid.FullName);
+                while (string.IsNullOrEmpty(map.EventsSection.Video))
+                {
+                    kvid = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
+                    vid = BeatmapDecoder.Decode(kvid.FullName);
+                }
+                map.EventsSection.Video = vid.EventsSection.Video;
+                map.EventsSection.VideoOffset = vid.EventsSection.VideoOffset;
+                //copy video
+                File.Copy(Path.Combine(Path.GetDirectoryName(kvid.FullName), vid.EventsSection.Video), Path.Combine("RandomizedMap", vid.EventsSection.Video));
+            }
+
+            //storyboard randomizer
+            /*
+            if (checkedListBox1.GetItemChecked(3))
+            {
+                var ksb = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
+                Beatmap sb = BeatmapDecoder.Decode(ksb.FullName);
+                while (!sb.EventsSection.Storyboard.)
+                {
+                    ksb = LoadedMaps.ElementAt(r.Next(0, LoadedMaps.Count));
+                    sb = BeatmapDecoder.Decode(ksb.FullName);
+                }
+                map.EventsSection.Video = sb.EventsSection.Video;
+                map.EventsSection.VideoOffset = sb.EventsSection.VideoOffset;
+                //copy video
+                File.Copy(Path.Combine(Path.GetDirectoryName(ksb.FullName), sb.EventsSection.Video), Path.Combine("RandomizedMap", sb.EventsSection.Video));
+            }
+            */
+
+            if(checkedListBox1.GetItemChecked(4))
+            {
+                map.DifficultySection.CircleSize = r.Next(0, 100) / 10f;
+            }
+
+            if (checkedListBox1.GetItemChecked(5))
+            {
+                map.DifficultySection.ApproachRate = r.Next(0, 100) / 10f;
+            }
+
+            if (checkedListBox1.GetItemChecked(6))
+            {
+                map.DifficultySection.OverallDifficulty = r.Next(0, 100) / 10f;
+            }
+
+            if (checkedListBox1.GetItemChecked(7))
+            {
+                map.DifficultySection.HPDrainRate = r.Next(0, 100) / 10f;
+            }
+
             map.MetadataSection.Artist =  "osu!map randomizer";
             map.MetadataSection.Title = "Randomized map";
             map.MetadataSection.Creator = "osu!map randomizer";
-            //create map directory
-            Directory.CreateDirectory("RandomizedMap");
-            //copy audio file
-            File.Copy(Path.Combine(Path.GetDirectoryName(ksong.FullName), song.GeneralSection.AudioFilename), Path.Combine("RandomizedMap", song.GeneralSection.AudioFilename));
+     
             //copy background
             File.Copy(Path.Combine(Path.GetDirectoryName(kmap.FullName), song.EventsSection.BackgroundImage), Path.Combine("RandomizedMap", song.EventsSection.BackgroundImage));
             map.Save(Path.Combine("RandomizedMap", "map.osu"));
+            ZipFile.CreateFromDirectory(Path.GetFullPath("RandomizedMap"), "RandomizedMap.osz");
+            Process.Start(Path.GetFullPath("RandomizedMap.osz"));
         }
     }
 }
